+++ packages/ide-extension/src/links/index.ts
@@ -1,1 +1,2 @@
 export { GuidedAnswersUriHandler } from './guided-answers-uri-handler';
+export { extractLinkInfo, generateExtensionLink, generateWebLink } from './link-info';